+++ packages/card_issuer/card_issuer.go
@@ -2,11 +2,12 @@ package card_issuer
 
 import (
 	"main.go/packages/interfaces"
+	"main.go/packages/issuer_tree"
 	"main.go/packages/types"
 )
 
 func GetCardIssuers(sequence []int, visitor interfaces.Visitor) ([]types.CardIssuer, error) {
-	singleInstance, error := GetInstance()
+	singleInstance, error := issuer_tree.GetInstance()
 	if error != nil {
 		return nil, error
 	}
